+++ R/sp.R
@@ -55,10 +55,13 @@
 #' }
 #' @export
 st_as_sf.Spatial = function(x, ...) {
-	if ("data" %in% slotNames(x))
+	if ("data" %in% slotNames(x)) {
+                if (!isTRUE(all.equal(row.names(x@data), row.names(x))))
+                    row.names(x@data) <- row.names(x)
 		df = x@data
-	else
+	} else {
 		df = data.frame(row.names = row.names(x)) # empty
+        }
 	if ("geometry" %in% names(df))
 		warning("column \"geometry\" will be overwritten by geometry column")
 	if (! requireNamespace("sp", quietly = TRUE))
@@ -301,8 +304,11 @@ as_Spatial = function(from, cast = TRUE, IDs = paste0("ID", seq_along(from))) {
 	if (inherits(from, "sf")) {
 		geom = st_geometry(from)
 		from[[attr(from, "sf_column")]] = NULL # remove sf column list
-		sp::addAttrToGeom(as_Spatial(geom, cast = cast, IDs = row.names(from)),
+		if (ncol(from))
+			sp::addAttrToGeom(as_Spatial(geom, cast = cast, IDs = row.names(from)),
 						  data.frame(from), match.ID = FALSE)
+		else
+			.as_Spatial(from, cast, IDs)
 	} else {
 		.as_Spatial(from, cast, IDs)
 	}