+++ R/stars.R
@@ -391,14 +391,15 @@ gdal_create = function(f, nxy, values, crs, xlim, ylim) {
 #' @param method character; method to create overview; one of: nearest, average, rms, gauss, cubic, cubicspline, lanczos, average_mp, average_magphase, mode
 #' @param layers integer; layers to create overviews for (default: all)
 #' @param options character; dataset opening options
+#' @param config_options named character vector with GDAL config options, like \code{c(option1=value1, option2=value2)}
 #' @param clean logical; if \code{TRUE} only remove overviews, do not add
 #' @param read_only logical; if \code{TRUE}, add overviews to another file with extension \code{.ovr} added to \code{file}
 #' @return \code{TRUE}, invisibly, on success
 #' @seealso \link{gdal_utils} for access to other gdal utilities that have a C API
 #' @export
 gdal_addo = function(file, overviews = c(2,4,8,16), method = "NEAREST", layers = integer(0), 
-					 options = character(0), clean = FALSE, read_only = FALSE) {
+					 options = character(0), config_options = character(0), clean = FALSE, read_only = FALSE) {
 	stopifnot(length(method) == 1, is.character(method), is.numeric(overviews))
 	invisible(CPL_gdaladdo(file, method, as.integer(overviews), as.integer(layers), as.character(options), 
-				 as.logical(clean)[1], as.logical(read_only)[1]))
+				 as.character(config_options), as.logical(clean)[1], as.logical(read_only)[1]))
 }